backend/app/config.py
from motor.motor_asyncio import AsyncIOMotorClient
from dotenv import load_dotenv
import os

# Load environment variables from .env file
load_dotenv()

# MongoDB configuration
MONGO_URI = os.getenv("MONGO_URI")
DATABASE_NAME = os.getenv("DATABASE_NAME")

#MongoDB connection 
client = AsyncIOMotorClient(MONGO_URI)
database = client[DATABASE_NAME]