backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
dnspython==2.7.0
fastapi==0.115.12
h11==0.14.0
idna==3.10
motor==3.7.0
pydantic==2.11.3
pydantic_core==2.33.1
pymongo==4.12.0
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
